Break up printer models
